pyproject.toml
[project]
name = "penny-finder-backend"
version = "0.1.0"
description = "Weekly briefing scraper for the Penny Finder app"
requires-python = ">=3.12"
dependencies = [
  "httpx>=0.27",
  "selectolax>=0.3.21",
  "feedparser>=6.0",
  "pydantic>=2.5",
  "python-dateutil>=2.8",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-asyncio>=0.23",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["scrapers*"]
